Make SSDLite more faithful to Tensorflow's
It's possible for the network to learn to cancel the BatchNormalization at the
end, but there isn't any reason to include it AFAIK, plus the Tensorflow
version doesn't seem to use it, so it has been removed.

The final pointwise convolution in the Tensorflow version does seem to use
biases though, so that has been included.